requirements.txt
amqp==5.1.1
asgiref==3.6.0
billiard==3.6.4.0
black==23.3.0
celery==5.2.7
certifi 
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
cron-descriptor==1.4.0
cryptography==40.0.1
Django==4.1.7
django-celery-beat==2.5.0
django-timezone-field==5.0
djangorestframework==3.14.0
exceptiongroup==1.1.1
Faker==18.4.0
idna==3.4
importlib-metadata==6.6.0
importlib-resources==5.12.0
iniconfig==2.0.0
kombu==5.2.4
libcurl-ct==7.84.0a2
load-dotenv==0.1.0
mypy-extensions==1.0.0
packaging==23.0
pathspec==0.11.1
pkg-about==1.0.8
platformdirs==3.2.0
pluggy==1.0.0
prompt-toolkit==3.0.38
pycparser==2.21
pytest==7.3.1
pytest-django==4.5.2
python-crontab==2.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
requests==2.28.2
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
uuid==1.30
vine==5.0.0
wcwidth==0.2.6
wincertstore==0.2
zipp==3.15.0
